fix: add wait compaction timeout
